# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AP Automation API",
    "description": "API for accounts payable automation with ERP-neutral extraction and configurable connectors",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard/packages": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Packages",
        "description": "Returns list of feedlot packages with optional filtering.",
        "operationId": "list_packages_dashboard_packages_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Period filter (e.g., '2025-11')",
              "title": "Period"
            },
            "description": "Period filter (e.g., '2025-11')"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PackageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by feedlot name or owner",
              "title": "Search"
            },
            "description": "Search by feedlot name or owner"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/models__api_responses__PackageSummary"
                  },
                  "title": "Response List Packages Dashboard Packages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Package Details",
        "description": "Returns complete package details including all invoices.",
        "operationId": "get_package_dashboard_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}/invoices": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Package Invoices",
        "description": "Returns all invoices in a package.",
        "operationId": "list_package_invoices_dashboard_packages__package_id__invoices_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by invoice status",
              "title": "Status"
            },
            "description": "Filter by invoice status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  },
                  "title": "Response List Package Invoices Dashboard Packages  Package Id  Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/packages": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Create Package",
        "description": "Create a new AP package.\n\nCreates an empty package ready to receive statement and invoice documents.",
        "operationId": "create_package_packages_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "List Packages",
        "description": "List AP packages with optional filtering.",
        "operationId": "list_packages_packages_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/api__routes__packages__PackageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "vendor_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Code"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Get Package",
        "description": "Get a specific AP package by ID.",
        "operationId": "get_package_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Packages"
        ],
        "summary": "Delete Package",
        "description": "Delete an AP package.",
        "operationId": "delete_package_packages__package_id__delete",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Package Packages  Package Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/summary": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Get Package Summary",
        "description": "Get processing summary for a package.",
        "operationId": "get_package_summary_packages__package_id__summary_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/statement": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Upload Statement",
        "description": "Upload a statement document to the package.",
        "operationId": "upload_statement_packages__package_id__statement_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_statement_packages__package_id__statement_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Statement Packages  Package Id  Statement Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/invoices": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Upload Invoices",
        "description": "Upload invoice documents to the package.",
        "operationId": "upload_invoices_packages__package_id__invoices_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_invoices_packages__package_id__invoices_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Invoices Packages  Package Id  Invoices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/extract": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Trigger Extraction",
        "description": "Trigger extraction workflow for a package.",
        "operationId": "trigger_extraction_packages__package_id__extract_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Extraction Packages  Package Id  Extract Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/reconcile": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Trigger Reconciliation",
        "description": "Trigger reconciliation for a package.",
        "operationId": "trigger_reconciliation_packages__package_id__reconcile_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Reconciliation Packages  Package Id  Reconcile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/post": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Post To Erp",
        "description": "Post package to ERP system.",
        "operationId": "post_to_erp_packages__package_id__post_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          },
          {
            "name": "connector_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "business_central",
              "title": "Connector Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post To Erp Packages  Package Id  Post Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_statement_packages__package_id__statement_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_statement_packages__package_id__statement_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "ready",
          "review",
          "blocked",
          "processing"
        ],
        "title": "InvoiceStatus",
        "description": "Invoice status values."
      },
...
      "PackageCreateRequest": {
        "properties": {
          "vendor_code": {
            "type": "string",
            "title": "Vendor Code",
            "description": "Vendor code or identifier"
          },
          "feedlot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedlot",
            "description": "Feedlot identifier"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "vendor_code"
        ],
        "title": "PackageCreateRequest",
        "description": "Request to create a new AP package."
      },
...
      "PackageDetailResponse": {
        "properties": {
          "header": {
            "$ref": "#/components/schemas/PackageDetailHeader"
          },
          "invoices": {
            "items": {
              "$ref": "#/components/schemas/InvoiceSummary"
            },
            "type": "array",
            "title": "Invoices"
          },
          "selected_invoice_detail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "header"
        ],
        "title": "PackageDetailResponse",
        "description": "Complete package detail response for the package drill-down view.\n\nIncludes package header, invoice list, and reconciliation summary."
      },
...
      "PackageStatus": {
        "type": "string",
        "enum": [
          "ready",
          "review",
          "blocked"
        ],
        "title": "PackageStatus",
        "description": "Package status values."
      },
...
      "api__routes__packages__PackageStatus": {
        "type": "string",
        "enum": [
          "pending",
          "extracting",
          "validating",
          "reconciling",
          "needs_review",
          "ready",
          "posting",
          "posted",
          "failed"
        ],
        "title": "PackageStatus",
        "description": "Status of an AP package."
      },
      "models__api_responses__PackageSummary": {
        "properties": {
          "package_id": {
            "type": "string",
            "title": "Package Id",
            "description": "Package identifier (e.g., PKG-2025-1130-531)"
          },
          "feedlot_name": {
            "type": "string",
            "title": "Feedlot Name",
            "description": "Feedlot display name"
          },
          "feedlot_code": {
            "type": "string",
            "title": "Feedlot Code",
            "description": "Short feedlot code (e.g., BF2, MC1)"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name",
            "description": "Cattle owner name"
          },
          "total_invoices": {
            "type": "integer",
            "title": "Total Invoices",
            "description": "Total invoices in package"
          },
          "total_lots": {
            "type": "integer",
            "title": "Total Lots",
            "description": "Total unique lots"
          },
          "ready_count": {
            "type": "integer",
            "title": "Ready Count",
            "description": "Invoices ready to post",
            "default": 0
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count",
            "description": "Invoices needing review",
            "default": 0
          },
          "blocked_count": {
            "type": "integer",
            "title": "Blocked Count",
            "description": "Invoices blocked",
            "default": 0
          },
          "total_dollars": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Dollars",
            "description": "Total package amount"
          },
          "status": {
            "$ref": "#/components/schemas/PackageStatus",
            "description": "Overall package status"
          },
          "statement_date": {
            "type": "string",
            "format": "date",
            "title": "Statement Date",
            "description": "Statement date"
          },
          "last_activity": {
            "type": "string",
            "title": "Last Activity",
            "description": "Relative time of last activity"
          },
          "last_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity At",
            "description": "Absolute timestamp of last activity"
          }
        },
        "type": "object",
        "required": [
          "package_id",
          "feedlot_name",
          "feedlot_code",
          "owner_name",
          "total_invoices",
          "total_lots",
          "total_dollars",
          "status",
          "statement_date",
          "last_activity",
          "last_activity_at"
        ],
        "title": "PackageSummary",
        "description": "Summary of a feedlot package for list display."
      }
    }
  }
}